[hotfix][test] Create a dedicated file for StreamTask cancellation tests
